+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds `FormSpacing` enum to DS exports.
+
 ## 0.24.0
 
 ### Breaking Changes