+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 0.24.1
+
 ### Adds
 
 - Adds `FormSpacing` enum to DS exports.